We have communication!
Did a lot of work on this today.  The Login function works now.  Basically the user enters a username and password and the login button checks to see if the credentials return the logged in version of the page or not.  It might be a hack, but I can't think of a better way to do it using basic http auth.

Now that I know the REST Client is working, I can also get the events from the server and parse them into our list.  Most of the code additions today were to get those lists working.